move a bunch of files around
